Handle unknown indicators.
