+++ bundles/statistics/statsgrid/view/search/IndicatorCollapse.jsx
@@ -1,45 +1,11 @@
 import React from 'react';
-import { Message, Collapse, CollapsePanel, Tooltip } from 'oskari-ui';
-import { IconButton } from 'oskari-ui/components/buttons';
-import { InfoCircleOutlined } from '@ant-design/icons';
+import { Message, Collapse, CollapsePanel } from 'oskari-ui';
+import { IndicatorRow } from './IndicatorRow';
 import styled from 'styled-components';
 
 const StyledCollapse = styled(Collapse)`
     margin-top: 20px;
 `;
-const Row = styled('div')`
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    word-break: break-all;
-`;
-const Actions = styled('div')`
-    display: flex;
-    flex-direction: row;
-    margin-left: 10px;
-`;
-const RemoveButton = styled(IconButton)`
-    margin-right: 10px;
-`;
-
-const getIndicatorText = (labels) => {
-    const { indicator, params, full } = labels;
-    let cutLength = 60;
-    let minLength = 20;
-    const dots = '... ';
-    if (indicator && full.length > cutLength) {
-        if (params) {
-            cutLength = cutLength - dots.length - params.length;
-            return indicator.substring(0, Math.max(minLength, cutLength)) + dots + params;
-        } else {
-            cutLength = cutLength - dots.length;
-            return indicator.substring(0, cutLength) + dots;
-        }
-    } else {
-        return full;
-    }
-};
 
 export const IndicatorCollapse = ({ state, controller }) => {
     return (
@@ -50,25 +16,9 @@ export const IndicatorCollapse = ({ state, controller }) => {
                 {state.indicators?.length < 1 ? (
                     <Message messageKey='indicatorList.emptyMsg' />
                 ) : (
-                    state.indicators?.map(indicator => {
-                        return (
-                            <Row key={indicator.indicator}>
-                                <Tooltip title={indicator.labels?.full}>
-                                    <span>{getIndicatorText(indicator.labels)}</span>
-                                </Tooltip>
-                                <Actions>
-                                    <RemoveButton
-                                        type='delete'
-                                        onClick={() => controller.removeIndicator(indicator)}
-                                    />
-                                    <IconButton
-                                        icon={<InfoCircleOutlined />}
-                                        onClick={() => controller.openMetadataPopup(indicator)}
-                                    />
-                                </Actions>
-                            </Row>
-                        );
-                    })
+                    state.indicators?.map(indicator => (
+                        <IndicatorRow key={indicator.indicator} indicator={indicator} controller={controller} />
+                    ))
                 )}
             </CollapsePanel>
         </StyledCollapse>
