+++ bundles/statistics/statsgrid2016/resources/locale/en.js
@@ -179,7 +179,8 @@ Oskari.registerLocalization({
             'onlyPartialDataForIndicators': 'Service did not return all data for {indicators, plural, one {the indicator} other {indicators}}'
         },
         'missing': {
-            'regionsetName': 'Unknown'
+            'regionsetName': 'Unknown',
+            'indicator': 'Unknown indicator'
         },
         'datacharts': {
             'flyout': 'Searched data',
